Move PackOptionsForm to separate file
The PackOptionsForm class was relocated from AnimationListControl.cs to its own file, PackOptionsForm.cs, to improve code organization and maintain compilation visibility.
